NLP/NLP/ours/source/pyproject.toml
[project]
name = "gsr-cacl"
version = "1.0.0"
description = "Graph-Structured Retrieval + Constraint-Aware Contrastive Learning for Financial Documents"
readme = "README.md"
requires-python = ">= 3.10"
authors = [{ name = "natmin322" }]
keywords = ["retrieval", "knowledge-graph", "financial", "rag", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core ML
    "torch>=2.0.0",
    "numpy>=1.24.0",
    # Transformer backbone (differentiable text encoder)
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
    "accelerate>=0.25.0",
    # LoRA fine-tuning
    "peft>=0.7.0",
    # Embeddings (for inference / benchmark compatibility)
    "langchain-core>=0.2.32",
    "langchain>=0.2.14",
    "langchain_huggingface>=0.0.3",
    "langchain-community>=0.2.16",
    # Retrieval
    "rank-bm25>=0.2",
    "faiss-cpu>=1.7.4",
    # Dataset / evaluation
    "datasets>=3.0.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.0.0",
    # Utilities
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "jupyterlab>=4.0",
    "ipywidgets>=8.1",
]
gpu = [
    "faiss-gpu>=1.7.2",
]

[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "SIM", "C4"]
ignore = [
    "E741",  # ambiguous variable name
    "N802",  # function name casing
    "N803",  # argument name casing
    "N806",  # variable in function casing
    "SIM102", # collapsible if
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
